Moved QueueStore to com.hz.collection
It belongs there since IQueue is also in that directory.
