+++ hazelcast-spring/src/test/java/com/hazelcast/spring/DummyQueueStore.java
@@ -16,7 +16,7 @@
 
 package com.hazelcast.spring;
 
-import com.hazelcast.core.QueueStore;
+import com.hazelcast.collection.QueueStore;
 
 import java.util.Collection;
 import java.util.Map;
