+++ hazelcast-spring/src/test/java/com/hazelcast/spring/DummyQueueStoreFactory.java
@@ -16,8 +16,8 @@
 
 package com.hazelcast.spring;
 
-import com.hazelcast.core.QueueStore;
-import com.hazelcast.core.QueueStoreFactory;
+import com.hazelcast.collection.QueueStore;
+import com.hazelcast.collection.QueueStoreFactory;
 
 import java.util.Properties;
