+++ hazelcast-spring/src/test/java/com/hazelcast/spring/TestFullApplicationContext.java
@@ -125,9 +125,9 @@ import com.hazelcast.core.MapStoreFactory;
 import com.hazelcast.cluster.Member;
 import com.hazelcast.cluster.MembershipListener;
 import com.hazelcast.multimap.MultiMap;
-import com.hazelcast.core.QueueStore;
-import com.hazelcast.core.QueueStoreFactory;
 import com.hazelcast.replicatedmap.ReplicatedMap;
+import com.hazelcast.collection.QueueStore;
+import com.hazelcast.collection.QueueStoreFactory;
 import com.hazelcast.ringbuffer.RingbufferStore;
 import com.hazelcast.ringbuffer.RingbufferStoreFactory;
 import com.hazelcast.crdt.pncounter.PNCounter;
