+++ hazelcast/src/main/java/com/hazelcast/collection/QueueStore.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.core;
+package com.hazelcast.collection;
 
 import java.util.Collection;
 import java.util.Map;
