+++ hazelcast/src/main/java/com/hazelcast/collection/QueueStoreFactory.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.core;
+package com.hazelcast.collection;
 
 import java.util.Properties;
